requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
scikit-image
opencv-python
matplotlib
seaborn
tqdm
pillow
shap
joblib
requests
